requirements.txt
anyio==3.7.0
asgiref==3.7.2
Django==4.2.1
graphql-core==3.2.3
idna==3.4
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
strawberry-graphql==0.178.0
typing_extensions==4.6.2
